+++ NEWS.md
@@ -1,3 +1,14 @@
+# MetaNet v0.1.3 Notes
+
+## Added
+
+- added `co_net_union` <2024-03-29, Fri>
+- added `params_list` argument for `c_net_plot`, easy to repeat calls to set default parameters <2024-03-28, Thu>
+
+## Fixed
+
+- fixed `zp_analyse()` as the Zi or Pi maybe NA <2024-04-09, Tue>
+
 # MetaNet v0.1.2 Notes
 
 ## Fixed
@@ -20,5 +31,5 @@
 
 - fixed R CMD check and Bioc check <2024-01-10>
 - submitted to CRAN <2024-01-10>
-- available on CRAN <https://cran.r-project.org/web/packages/MetaNet/> <2024-01-12>
+- available on CRAN <https://CRAN.R-project.org/package=MetaNet> <2024-01-12>